refactor: extract wallet reset saga and trigger it from an action
