+++ ts/features/itwallet/common/saga/index.ts
@@ -6,11 +6,13 @@ import { checkWalletInstanceStateSaga } from "../../lifecycle/saga/checkWalletIn
 import { handleWalletCredentialsRehydration } from "../../credentials/saga/handleWalletCredentialsRehydration";
 import { itwTrialId } from "../../../../config";
 import { itwCieIsSupported } from "../../identification/store/actions";
+import { watchItwLifecycleSaga } from "../../lifecycle/saga";
 
 export function* watchItwSaga(): SagaIterator {
   yield* fork(checkWalletInstanceStateSaga);
   yield* fork(handleWalletCredentialsRehydration);
   yield* fork(watchItwIdentificationSaga);
+  yield* fork(watchItwLifecycleSaga);
 
   // TODO: [SIW-1404] remove this CIE check and move the logic to xstate
   yield* put(itwCieIsSupported.request());
