+++ ts/features/itwallet/lifecycle/saga/checkWalletInstanceStateSaga.ts
@@ -1,45 +1,14 @@
 import { Errors } from "@pagopa/io-react-native-wallet";
-import { deleteKey } from "@pagopa/io-react-native-crypto";
-import { all, call, put, select } from "typed-redux-saga/macro";
-import { identity, pipe } from "fp-ts/lib/function";
+import { call, select } from "typed-redux-saga/macro";
 import * as O from "fp-ts/lib/Option";
-import * as RA from "fp-ts/lib/ReadonlyArray";
 import { getAttestation } from "../../common/utils/itwAttestationUtils";
 import { ensureIntegrityServiceIsReady } from "../../common/utils/itwIntegrityUtils";
 import { itwIntegrityKeyTagSelector } from "../../issuance/store/selectors";
 import { ReduxSagaEffect } from "../../../../types/utils";
 import { itwLifecycleIsOperationalOrValid } from "../store/selectors";
-import { itwLifecycleWalletReset } from "../store/actions";
-import { walletRemoveCardsByType } from "../../../newWallet/store/actions/cards";
 import { sessionTokenSelector } from "../../../../store/reducers/authentication";
 import { assert } from "../../../../utils/assert";
-import { itwCredentialsSelector } from "../../credentials/store/selectors";
-import { StoredCredential } from "../../common/utils/itwTypesUtils";
-import { isIos } from "../../../../utils/platform";
-
-const getKeyTag = (credential: O.Option<StoredCredential>) =>
-  pipe(
-    credential,
-    O.map(x => x.keyTag)
-  );
-
-export function* handleWalletInstanceReset(integrityKeyTag: string) {
-  const { eid, credentials } = yield* select(itwCredentialsSelector);
-  yield* put(itwLifecycleWalletReset());
-  yield* put(walletRemoveCardsByType("itw"));
-
-  // Remove all keys within the wallet.
-  // On iOS skip the integrity key tag as it is managed by the App Attest service.
-  const itwKeyTags = pipe(
-    [
-      isIos ? O.none : O.of(integrityKeyTag),
-      getKeyTag(eid),
-      ...credentials.map(getKeyTag)
-    ],
-    RA.filterMap(identity)
-  );
-  yield* all(itwKeyTags.map(deleteKey));
-}
+import { handleWalletInstanceReset } from "./handleWalletInstanceResetSaga";
 
 export function* getAttestationOrResetWalletInstance(integrityKeyTag: string) {
   const sessionToken = yield* select(sessionTokenSelector);
@@ -54,7 +23,7 @@ export function* getAttestationOrResetWalletInstance(integrityKeyTag: string) {
       err instanceof Errors.WalletInstanceRevokedError ||
       err instanceof Errors.WalletInstanceNotFoundError
     ) {
-      yield* call(handleWalletInstanceReset, integrityKeyTag);
+      yield* call(handleWalletInstanceReset);
     }
   }
 }
